Add parent field to node to help with split()
During an insertion, it is possible that the items array we wish to
insert into is full. When this is the case, the items array must be
split such that:
1. A separation value (median of items and item to be inserted) is
inserted into parent node's items.
2. Values to left and right of separation value are put into 2 new
nodes which are the children of separation node.
Note that this process may need to be repeated multiple times for
parent, their parent, etc.

The 'parent' field should help simplify this process.