Refactor split method to reuse current node
Currently, split() splits a node by creating a left and right node and
copying values from the current node to these nodes. The current node is
then left to be garbage collected.

This commit improves split() by reusing the current node as the left
node. By reusing the current node, we skip unnecessary garbage
collection, we avoid unnecessary copying, and we shorten the split()
method.

To do this, this commit also adds truncate() methods for both items and
split. These methods give nodes a simple way to alter the size of their
'items' and 'children' fields.

NOTE: I've also slightly altered the newNode() function. The 'items'
and 'children' field should be allowed to be nil, so I've removed the
calls to make when these arguments are nil.