Refactor how Search() does equality checks
Currently, Search() performs up to 2 find operations in each iteration:
1. findEqual(), which contains internal call to find()
2. find()

Due to how find() works, this is hugely unnecessary. The find() function
returns the smallest index where:
  itemSearchedFor < items[index]
As such, any equal items will necessarily be located at index - 1.
Given that this is the case, I've decided to replace findEqual() with
match(), a method which is only used to compare items. Search now only
needs to call find() once, as the index returned by find() can be used
for both checking equality and for picking a child pointer to follow for
the next iteration.