Replace slice nil checks with len comparisons
Currently, there are a few places in the code (usually when checking if
current node is a leaf) where a slice is determined to be empty if
slice == nil. I felt that the intent of these statements wasn't
especially clear, so I've replaced them with len(slice) == 0.

In doing so, I realized that a few of these checks were redundant. In
particular, all of the find / search functions don't need to explicitly
test that len(node.children) == 0 to determine that a node is a leaf. In
these cases, I've removed the check entirely.