Add node method to get pos in parent's children
The node.nthChildOfParent() method allows a node to determines its
index in node.parent.children. This is useful during rebalances, as this
information is needed to determine a node's siblings and separator
items.

Currently, this is done by performing a linear search of
node.parent.children. nthChildOfParent() improves on this by deferring
to n.items.find(n.items[0]) (which uses a binary search) if n.items is
not empty. Due to how find() is structured, and given that n.items[0] is
guaranteed to not equal any item in n.parent.items, the index returned
by find() will be the same as the aforementioned linear search.

In making this change, I've also realized that much of the work done by
the node.leftSibling() and node.rightSibling() methods is unnecessary.
As such, I've removed these methods entirely. Siblings are now found
using node.nthChildOfParent() as part of rebalance().