Reconcieve Iterator as unidirectional
Initially, btree.Iterator was conceived as being bidirectional - able to
move forwards and backwards through a tree. The Iterator would have had
separate Next() and Prev() methods to iterate in different directions.
However, after finishing Next() and beginning work on Prev(), I realized
that creating such an Iterator would have required storing a lot of
state data. It would have been double what I initially intended - one
set for the next or previous element and one set for last visited
element. The Iterator would also have needed to track the direction of
the direction of iteration to know when it had changed.

Because this was so much uglier than I had anticipated, and because I
realized that most users would only want to iterate in one direction,
I've decided to rethink Iterator.

Iterator now has:
1. Different properties (and fewer properties overall)
   Instead of separate 'next' and 'prev' fields, Iterator now has a
   single 'curr' field to track the node being iterated over. Iterator
   also now has a 'dir' field, a value of either 1 or -1 which describes
   how Next() should iterate.
2. Two constructors - NewIterator and NewReverseIterator
   An Iterator's direction is set and fixed at creation.
3. Only HasNext() and Next() methods
   There is no longer separate 'HasPrev' or 'Prev' methods.
   Instead, the notion of 'next' is tied to an iterators direction.

On the whole, I feel this new design is superior in every way to what I
had before.