Add function to quickly make BTree from item array
The Bulkload function allows users to quickly create a BTree using
existing, ordered data. Because the data is ordered, the Bulkload
function is able to skip the most time-consuming part of Insert -
finding the node in which to insert the item. Instead, new items will
always be inserted into the maximum (i.e. rightmost) node.

While Bulkload is extremely fast, their are a few serious caveats for
using it:
1. The array of items must be sorted.
2. The array of items may not contain duplicates.
The functions behavior is not defined for either of these cases.

NOTE: Because Bulkload creates a tree rather than acting on an existing
tree, I've decided to make it a function rather than a method as it was
before.