Fix Iterator bug involving excess iterations
Currently, fully iterating Iterator will result in one too many
iterations. This is due to how I've designed the Next() method. When
iterating through a leaf node's items, each call to next increments
iterator.itemIndex to point to the next item. This is problematic, as it
doesn't address what should happen when iterator.itemIndex > len(items).
Instead, that case is handled in the next call to Next(). Because of
this, iterating over n items requires n+1 calls to Next(), with the last
call setting iterator.curr to nil and hence causing HasNext() to return
false.

To fix this, I've changed the Iterator so that the next value is always
known and is always given by iterator.curr[iterator.itemIndex]. This
required making the following changes:
1. Value to return is defined at beginning of Next()
   As noted above, the next value is always given by
   iterator.curr[iterator.itemIndex]. These leaves the rest of the
   function to focus solely on updating iterator.curr,
   iterator.itemIndex, and iterator.childIndex for subsequent calls to
   Next().
2. Iterator now begins at min or max nodes rather than root
   Having root as initial value of 'curr' caused some design issues, as
   it meant that the first call to Next() needed to be treated
   differently than subsequent calls (next and next-next had to be found
   during the same call). By using min and max for normal and reverse
   iteration respectively, we avoid this issue entirely.
3. Check and handle leaf nodes if they are full during same call
   To fix the issues mentioned above, the Iterator's properties (curr,
   itemIndex, and childIndex) will be updated in the same call that the
   current leaf node is found to be full.