Cleanup tests and improve benchmarks
I've made the following changes to the test file:
1. Added uniqueInputsN() and duplicateInputsN() to generate randomly
   ordered slices of Items. These functions replace the setup process
   currently used in most of the tests.
2. Moved rand seeding to init(), as it only needs to happen once.
3. Created multiple Benchmark functions for Iterator (forward and
   reverse).
4. Expanded test cases for various of the test functions.

In addition, I've fixed a bug in Iterator. Currently, iterator.Next()
will fail if the tree being iterated over is empty. This happens because
iterator.HasNext() only checks that curr (which would be the root node)
is not nil. To account for this, HasNext() now also checks that
len(curr.items) is not 0.